+++ src/Router/Router.fs
@@ -1,27 +1,76 @@
 namespace KafkaApplication.Router
 
-module internal ContentBasedRouter =
-    open Kafka
-    open KafkaApplication
-    open OptionOperators
+open Kafka
 
-    let private formatLogMessage (EventName eventName) (streamName: StreamName) =
-        (sprintf "Route event %s to %A ..." eventName streamName)
+// Router
 
-    let private sendToStream log produceTo eventToRoute processedBy (stream: StreamName) =
-        log <| sprintf "-> Sending to stream %A ..." stream
+[<RequireQualifiedAccess>]
+module internal Router =
+    [<RequireQualifiedAccess>]
+    module Configuration =
+        open FSharp.Data
+        open System.IO
+        open ServiceIdentification
 
-        eventToRoute
-        |> EventToRoute.route processedBy
-        |> produceTo stream
+        type private RoutingSchema = JsonProvider<"src/Router/schema/routingSchema.json">
 
-    let routeEvent log produce processedBy router eventToRoute =
-        let eventType = eventToRoute.Raw.Event
+        let parse path =
+            result {
+                let rawRouting =
+                    path
+                    |> File.ReadAllText
+                    |> RoutingSchema.Parse
 
-        router
-        |> Router.getStreamFor eventType
-        |>! (
-            tee (formatLogMessage eventType >> log)
-            >> tee (sendToStream log produce eventToRoute processedBy)
-            >> ignore
-        )
+                let! routing =
+                    rawRouting.Route
+                    |> Seq.map (fun route ->
+                        result {
+                            let! topicInstance =
+                                route.TargetStream
+                                |> Instance.parse "-"
+                                |> Result.ofOption (StreamNameIsNotInstance route.TargetStream)
+
+                            return (EventName route.Event, StreamName.Instance topicInstance)
+                        }
+                    )
+                    |> Seq.toList
+                    |> Result.sequence
+
+                return
+                    routing
+                    |> Map.ofList
+                    |> RouterConfiguration
+            }
+
+        let getStreamFor event (RouterConfiguration router) =
+            router
+            |> Map.tryFind event
+
+        let getOutputStreams (RouterConfiguration router) =
+            router
+            |> Map.toList
+            |> List.map snd
+
+    [<RequireQualifiedAccess>]
+    module Routing =
+        open KafkaApplication
+        open global.Option.Operators
+
+        let private formatLogMessage (EventName eventName) (streamName: StreamName) =
+            sprintf "Route event %s to %A ..." eventName streamName
+
+        let private sendToStream log produceTo eventToRoute (stream: StreamName) =
+            log <| sprintf "-> Sending to stream %A ..." stream
+
+            eventToRoute
+            |> produceTo stream
+
+        let routeEvent log getEventType produce router eventToRoute =
+            let eventType = eventToRoute |> getEventType
+
+            router
+            |> Configuration.getStreamFor eventType
+            |>! (
+                tee (formatLogMessage eventType >> log)
+                >> sendToStream log produce eventToRoute
+            )